requirements.txt
playwright>=1.41.0
openpyxl>=3.1.2
pandas>=2.1.4
python-dotenv>=1.0.0
jsonschema>=4.20.0
Pillow>=10.2.0
